+++ databricks/koalas/tests/data_type_ops/test_num_ops.py
@@ -16,52 +16,13 @@
 
 import datetime
 import numpy as np
-import pandas as pd
-import decimal
 
-from databricks import koalas as ks
 from databricks.koalas.config import option_context
+from databricks.koalas.data_type_ops.testing_utils import TestCasesUtils
 from databricks.koalas.testing.utils import ReusedSQLTestCase
 
 
-class NumOpsTest(ReusedSQLTestCase):
-    @property
-    def numeric_psers(self):
-        dtypes = [np.float32, float, int, np.int32]
-        sers = [pd.Series([1, 2, 3], dtype=dtype) for dtype in dtypes]
-        sers.append(pd.Series([decimal.Decimal(1), decimal.Decimal(2), decimal.Decimal(3)]))
-        return sers
-
-    @property
-    def numeric_ksers(self):
-        return [ks.from_pandas(pser) for pser in self.numeric_psers]
-
-    @property
-    def non_numeric_psers(self):
-        psers = {
-            "string": pd.Series(["x", "y", "z"]),
-            "datetime": pd.to_datetime(pd.Series([1, 2, 3])),
-            "bool": pd.Series([True, True, False]),
-            "date": pd.Series(
-                [datetime.date(1994, 1, 1), datetime.date(1994, 1, 2), datetime.date(1994, 1, 3)]
-            ),
-            "categorical": pd.Series(["a", "b", "a"], dtype="category"),
-        }
-        return psers
-
-    @property
-    def non_numeric_ksers(self):
-        ksers = {}
-
-        for k, v in self.non_numeric_psers.items():
-
-            ksers[k] = ks.from_pandas(v)
-        return ksers
-
-    @property
-    def numeric_pser_kser_pairs(self):
-        return zip(self.numeric_psers, self.numeric_ksers)
-
+class NumOpsTest(ReusedSQLTestCase, TestCasesUtils):
     def test_add(self):
         for pser, kser in self.numeric_pser_kser_pairs:
             self.assert_eq(pser + pser, kser + kser)
